refactor(web): Extract toUtcDate into a shared date util
Move the naive-timestamp-as-UTC parser out of session.mapping into
shared/utils/dates/utc-date.ts and reuse it in plan.mapping and the progress
chart, which previously parsed backend timestamps with new Date() and so read
them as local time.
